[Feature-0523] - {Positions} 직급 수정 #134

+++ src/pages/members/Positions.js
@@ -1,4 +1,4 @@
-import { callPositionsAPI, callAddPositionAPI, callDeletePositionAPI } from "../../apis/MemberAPICalls"; // addPosition 추가
+import { callPositionsAPI, callAddPositionAPI, callDeletePositionAPI, callUpdatePositionStatusAPI,callAllPositionsAPI } from "../../apis/MemberAPICalls"; // addPosition 추가
 import { useSelector, useDispatch } from "react-redux";
 import { useEffect, useState } from "react";
 import { Table, ToggleButtonGroup } from "react-bootstrap";
@@ -12,15 +12,17 @@ function Positions() {
 
     const dispatch = useDispatch();
     const positionList = useSelector(state => state.positionReducer);
+  
     const [positionDialogOpen, setPositionDialogOpen] = useState(false);
     const [selectedPosition, setSelectedPosition] = useState(null);
     const [selectedPositionNos, setSelectedPositionNos] = useState({});
     const [isEditMode, setIsEditMode] = useState(false);
     const [newPosition, setNewPosition] = useState({ positionName: "", positionLevel: "", authority: "", positionStatus: "Y" });
 
-
+    console.log(useSelector(state => state.positionReducer));
+    
     useEffect(() => {
-        dispatch(callPositionsAPI());
+        dispatch(callAllPositionsAPI());
     }, []);
 
 
@@ -60,18 +62,26 @@ function Positions() {
     const addPosition = async () => {
         await dispatch(callAddPositionAPI(newPosition));
         setIsEditMode(false);
-        await dispatch(callPositionsAPI());
+        await dispatch(callAllPositionsAPI());
     }
 
     const cancelEdit = () => {
         setIsEditMode(false);
         setNewPosition({ positionName: "", positionLevel: "", authority: "" });
+     dispatch(callAllPositionsAPI());
+       
     }
 
     const deletePosition = async (positionNo) => {
-        await dispatch(callDeletePositionAPI(positionNo));
-        await dispatch(callPositionsAPI());
+        const confirmed = window.confirm('해당직급 사원이 없을경우에만 삭제 가능합니다.\n정말 삭제하시겠습니까? '); // 사용자에게 확인 다이얼로그 표시
+    
+        if (confirmed) { // 사용자가 확인을 눌렀을 때만 삭제 진행
+            await dispatch(callDeletePositionAPI(positionNo));
+            await dispatch(callAllPositionsAPI());
+        }
     }
+
+    
 
 
 
@@ -84,7 +94,6 @@ function Positions() {
                     <ButtonGroup buttons={[{ label: '등록', styleClass: 'move', onClick: handleEditModeToggle }]} />
                     <br />
                     <br />
-
                     <Table>
                         <thead>
                             <tr>
@@ -103,15 +112,21 @@ function Positions() {
                                     <td>
                                         <Radio
                                             name={`positionStatus-${position?.positionNo}`}
-                                            checked={selectedPositionNos[position?.positionNo] === false}
-                                            onClick={() => handleRadioClick(position?.positionNo, false)} // 누르는 버튼에 따라 false 전달
+                                            checked={position.positionStatus === 'N'}
+                                            onClick={
+                                                () => dispatch(callUpdatePositionStatusAPI(position.positionNo))
+                                                .then(() => {dispatch(callAllPositionsAPI())})
+                                            } // 누르는 버튼에 따라 false 전달
                                         />
                                     </td>
                                     <td>
                                         <Radio
                                             name={`positionStatus-${position?.positionNo}`}
-                                            checked={selectedPositionNos[position?.positionNo] === true}
-                                            onClick={() => handleRadioClick(position?.positionNo, true)} // 누르는 버튼에 따라 true 전달
+                                            checked={position.positionStatus === 'Y'}
+                                            onClick={
+                                                () => dispatch(callUpdatePositionStatusAPI(position.positionNo))
+                                                .then(() => {dispatch(callAllPositionsAPI())})
+                                            } // 누르는 버튼에 따라 true 전달
                                         />
                                     </td>
                                     <td>{position && position.positionName}</td>
@@ -126,7 +141,7 @@ function Positions() {
                     </Table>
                     <Dialog open={positionDialogOpen} onClose={onDialogClickHandler}>
                         <DialogTitle>직급 수정</DialogTitle>
-                        <PositionForm position={selectedPosition} />
+                        <PositionForm position={selectedPosition} onDialogClickHandler={onDialogClickHandler} />
                     </Dialog>
                     {isEditMode &&
                         <div>
